Move EventDialogSectionProps to EventDialog.types.ts
