+++ packages/x-scheduler/src/internals/components/event-dialog/EventDialog.types.ts
@@ -2,6 +2,7 @@ import type * as React from 'react';
 import type { SchedulerRenderableEventOccurrence } from '@mui/x-scheduler-internals/models';
 import type { DialogProps } from '@mui/material/Dialog';
 import type { EventDialogOptionalRenderers } from './EventDialogOptionalRenderersContext';
+import type { ControlledValue } from './utils';
 
 export interface EventDialogProps extends DialogProps {
   /**
@@ -29,4 +30,16 @@ export interface EventDialogProviderProps {
 export interface EventDialogTriggerProps extends React.HTMLAttributes<HTMLElement> {
   occurrence: SchedulerRenderableEventOccurrence;
   children: React.ReactNode;
+}
+
+/**
+ * Props shared by the General tab section components.
+ * Temporary contract until the form context replaces the prop plumbing (#22868).
+ */
+export interface EventDialogSectionProps {
+  occurrence: SchedulerRenderableEventOccurrence;
+  controlled: ControlledValue;
+  setControlled: React.Dispatch<React.SetStateAction<ControlledValue>>;
+  errors: Record<string, string | string[]>;
+  setErrors: (errors: Record<string, string | string[]>) => void;
 }
